assembler: Refactor module structure
