+++ assembler/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod scanner;
+pub mod models;
 pub mod error;
 pub mod parser;
