+++ assembler/src/models.rs
@@ -1,0 +1,5 @@
+pub mod span;
+pub mod token;
+
+pub use span::{Location, Span};
+pub use token::{Token, TokenWithSpan};
