+++ assembler/src/models/token/span.rs
@@ -1,0 +1,10 @@
+use crate::models::Span;
+
+use super::Token;
+
+#[derive(Debug)]
+#[allow(clippy::module_name_repetitions)]
+pub struct TokenWithSpan {
+    pub token: Token,
+    pub span: Span,
+}
